AB#587 moved verify function and global flags to own file
